Devs

+++ street-goods-burger-backend/db/migrate/20220610134907_create_favorite_foods.rb
@@ -2,7 +2,7 @@ class CreateFavoriteFoods < ActiveRecord::Migration[7.0]
   def change
     create_table :favorite_foods do |t|
       t.string :name, null: false, default: ''
-      t.bigint :price, null: false, default: 0.00
+      t.decimal :price, null: false, default: 0.00
       t.string :category, null: false, default: ''
       t.string :description, null: false, default: ''
       t.integer :like, null: false, default: 0
